Change mbedtls configuration to only support TLS_PSK_WITH_AES_128_CCM_8 which is required for LwM2M 1.0.2
The LwM2M 1.0.2 specification states that
"A LwM2M Server MUST support the Pre-Shared Key mode of DTLS with the following ciphersuites:"
- TLS_PSK_WITH_AES_128_CCM_8, as defined in [RFC6655]
- TLS_PSK_WITH_AES_128_CBC_SHA256, as defined in [RFC5487]
A LwM2M Client MUST support the Pre-Shared Key mode of DTLS with at least one of the ciphersuites specified for the LwM2M Server."

Changing the default configuration like that makes a whole lot of
wrapper code invalid which has been commented out then. And the
mbedtls-platform-support as a whole is removed.